Save modified images by their uuid rather than their modifications
We hash the modifications to be able to filter on them.

Notice the relative smallness of this change, due to the refactoring
done earlier, which allows for greater flexibility with less code.

Displaying paintovers seems to be broken, no requests going to the
server to even request the paintovers, not sure what's going on.

git-svn-id: https://graphics.tudelft.nl/svn/BachCASAM@660 9b22acdf-97ab-464f-81e2-08fcc4a6931f